Fixes untransferred code
